# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Экскурсии по Удмуртской Республике",
        "description": "API документация для работы с сервисом экскурсий по УР",
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/user/signup": {
            "post": {
                "tags": [
                    "Пользователь"
                ],
                "summary": "Зарегистрировать нового пользователя",
                "description": "Создаёт и возвращает нового пользователя в случае успеха",
                "operationId": "0e3c7d64c2fa728e7deb9063d9af33b3",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "required": [
                                            "first_name",
                                            "last_name",
                                            "birthday",
                                            "email",
                                            "password"
                                        ],
                                        "properties": {
                                            "first_name": {
                                                "type": "string",
                                                "example": "Иван"
                                            },
                                            "last_name": {
                                                "type": "string",
                                                "example": "Иванов"
                                            },
                                            "patronymic": {
                                                "type": "string",
                                                "example": "Иванович"
                                            },
                                            "birthday": {
                                                "type": "string",
                                                "example": "01.01.2000"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "foo@bar.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "password"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная регистрация",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User created"
                                        },
                                        "data": {
                                            "properties": {
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "Иван"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Иванов"
                                                },
                                                "patronymic": {
                                                    "type": "string",
                                                    "example": "Иванович"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "foo@bar.com"
                                                },
                                                "birthday": {
                                                    "type": "string",
                                                    "example": "2000-01-01T00:00:00.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2000-01-01T00:00:00.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2000-01-01T00:00:00.000000Z"
                                                },
                                                "id": {
                                                    "type": "number",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации данных",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "properties": {
                                                "first_name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The first name field is required."
                                                    ]
                                                },
                                                "last_name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The last name field is required."
                                                    ]
                                                },
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The email field is required."
                                                    ]
                                                },
                                                "password": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The password field is required"
                                                    ]
                                                },
                                                "birthday": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The birthday field is required"
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/signin": {
            "post": {
                "tags": [
                    "Пользователь"
                ],
                "summary": "Войти в аккаунт",
                "description": "Проверяет credentials и создаёт новый токен в случае успеха",
                "operationId": "462c3893ea1ca41948fc2ab2ef3a0209",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "required": [
                                            "email",
                                            "password"
                                        ],
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "example": "foo@bar.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "password"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешный вход",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logged in"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неверные credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Wrong credentials"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}